switching to i32

+++ src/data/decode.rs
@@ -18,9 +18,9 @@ pub fn decode_png(path: std::path::PathBuf) -> Vec<f64> {
 
 pub fn shuffle_dataset(
     x: &mut Vec<Vec<f64>>,
-    y: &mut Vec<f64>,
+    y: &mut Vec<i32>,
     rng: &mut Rng,
-) -> (Vec<Vec<f64>>, Vec<f64>) {
+) -> (Vec<Vec<f64>>, Vec<i32>) {
     for i in (1..y.len()).rev() {
         let j = (rng.next_f64() * (i + 1) as f64) as usize;
         x.swap(i, j);
